Fix miscellaneous pricing bugs (#16)
* Fix mrap class names in prices

* Fix plasma price inconsistencies

The price of blood supplements already depends on mass (line 37). These if statements made it depend on mass twice. Also it only ever affected the price of plasma, because blood and saline have already exited from the function.

* Add price for laser designator

The exact price may change later. The important part is that it now has a price. Previously it was sold for the overthrow default $10.

* Fix empty rope sold in hardware stores

The base class ACE_ropeBase was being sold even though it is not a real item.

* Fix inconsistent rope prices

* Fix ropes and UAV terminal not showing up in stores