Improve shop item discovery
Previously only direct children of certain 5 classes and their children (2 levels deep) were searched. Now all items are searched.
In vanilla Arma 3 the only difference is that binoculars and NVGs can now be found in shop inventories. With mods such as RHS however, this may have a huge effect on shop inventory. This may also affect the prices of a few items that were previously searched but not added to shops.