Refactor item info finding code in dialogs
This commit is only refactoring, behavior of the code has not changed. The code that finds pictures, names and descriptions for items displayed in UI dialogs has been moved into a separate function and it is called from almost every dialog. This new code will probably run as fast or slightly faster than the old. Additionally there is some optimization and cleanup of dead code.