Fix race condition when deleting vehicle crew
The deleteVehicleCrew is hard to use correctly. It must be executed with remoteExec on the computer where the vehicle is local, but this causes a delay before the code is really executed. During this delay, the unit might have disembarked or the vehicle might have been deleted, causing deleteVehicleCrew to fail. A real world example was in fn_cleanup.sqf where the vehicle was actually deleted right after calling remoteExec. Because of this, around 70% of time the crew was not properly deleted.