Fix helicopters not cleaned on landing
Sometimes helicopters land just briefly and take off again, so we need to check for the landing status every second instead of 10 seconds. This does not cause much lag because this check is only executed when the helicopter approaches the landing spot. There are also better ways to check if a helicopter has landed that may be implemented later.