Replace special case SHK_pos calls with getPos
These cases all had a random lower bound for the distance in the SHK_pos calls. That makes them a bit harder to convert to getPos. Conceptually, having a random lower bound from 0 to 100 means that the sampling density for the random positions is uniform for the first 100 meters and then starts to decrease with distance. Without that random lower bound the sampling density starts to decrease immediately from 0 (see the "wrong" picture in http://www.anderswallin.net/2009/05/uniform-random-points-in-a-circle-using-polar-coordinates/).

In the case of fugitive.sqf and informant.sqf, the result of the call doesn't matter that much, so the random lower bound was just removed. In getRandomRoadPosition, the uniform density of the first 100m might actually matter, so it was implemented properly using getPos.