Add proper Malden flags
Create new object, a flagpole with the Malden flag in it. This flagpole will be used for captured bases. Also fix the wrong color and aspect ratio of the existing Malden flag marker.

The flagpole texture is 512x256 and features wrinkles, like other Arma flags. The marker texture is 128x128 and has transparent areas above and below, like other Arma flags.