Fix remains of an ancient failed search && replace
It seems that long ago in commits 25ccb33d and 61aa9262 all and-words were replaced with && and all or-words with ||. That replace action failed in that also words in comments and strings were replaced. There were still some remains of that, which are now fixed.

The description string of police station is also converted to structured text where && was causing errors.