Add fallback NATO faction
If there were no vehicles in the main NATO faction, they are taken from the fallfack faction. This fixes an issue with Livonia where the woodland NATO faction has no vehicles, so they should be taken from the tropic NATO faction. In the future this can be further expanded to infantry and groups as well.