Fix locality issues with looting items.
Looting scripts are being called for bodies running on multiple machines, but they used a lot of commands with local arguments. I'm not sure what kind of bugs this has caused but they are now fixed. The "remove*WeaponItems" commands were just removed because the weapon is going to be deleted anyway.