Optimize body search code when looting
This does not change the looting results, simply optimizes performance. In my tests if all the bodies in the world are inside the looting radius, this new code is a bit slower, but it becomes faster whenever over half of the bodies are outside the looting radius, which is usually the case in Overthrow.